@@ -42,10 +42,8 @@ namespace DigitalRuby.IPBan
         /// <returns>Task</returns>
         public static async Task Main(string[] args)
         {
-            if (args.Length != 0 && (args[0].Equals("info", StringComparison.OrdinalIgnoreCase) ||
-                args[0].Equals("-info", StringComparison.OrdinalIgnoreCase)))
+            if (ProcessCommandLine(args))
             {
-                Logger.Warn("System info: {0}", OSUtility.OSString());
                 return;
             }
 
@@ -62,6 +60,32 @@ namespace DigitalRuby.IPBan
                 service?.Dispose();
                 return Task.CompletedTask;
             });
+        }
+
+        private static bool ProcessCommandLine(string[] args)
+        {
+            if (args.Length != 0)
+            {
+                if (args[0].Contains("info", StringComparison.OrdinalIgnoreCase))
+                {
+                    Logger.Warn("System info: {0}", OSUtility.OSString());
+                    return true;
+                }
+                else if (args[0].Contains("logfiletest", StringComparison.OrdinalIgnoreCase))
+                {
+                    if (args.Length != 6)
+                    {
+                        Console.WriteLine("Usage: log-filename regex-failure-filename regex-failure-timestamp-format regex-success-filename regex-success-timestamp-format");
+                        Console.WriteLine("Can use a . to not specify the regex or timestamp format");
+                    }
+                    else
+                    {
+                        IPBanLogFileTester.RunLogFileTest(args[1], args[2], args[3], args[4], args[5]);
+                    }
+                    return true;
+                }
+            }
+            return false;
         }
     }
 }
