@@ -290,12 +290,12 @@ namespace DigitalRuby.IPBanCore
                 ipAddressObj = ipAddressObj.Clean();
                 byte[] ipBytes = ipAddressObj.GetAddressBytes();
                 SqliteDBTransaction tran = transaction as SqliteDBTransaction;
-                using SqliteDataReader reader = ExecuteReader("SELECT IPAddressText, LastFailedLogin, FailedLoginCount, BanDate," +
+                using SqliteDataReaderWrapper reader = ExecuteReader("SELECT IPAddressText, LastFailedLogin, FailedLoginCount, BanDate," +
                     "State, BanEndDate, UserName, Source FROM IPAddresses WHERE IPAddress = @Param0",
-                    tran?.DBConnection, tran?.DBTransaction, ipBytes);
-                if (reader.Read())
+                    tran?.DBConnection, tran?.DBTransaction, false, ipBytes);
+                if (reader.Reader.Read())
                 {
-                    entry = ParseIPAddressEntry(reader);
+                    entry = ParseIPAddressEntry(reader.Reader);
                     return true;
                 }
             }
@@ -339,13 +339,13 @@ namespace DigitalRuby.IPBanCore
                 ipAddressObj = ipAddressObj.Clean();
                 byte[] ipBytes = ipAddressObj.GetAddressBytes();
                 SqliteDBTransaction tran = transaction as SqliteDBTransaction;
-                using SqliteDataReader reader = ExecuteReader("SELECT BanDate, BanEndDate FROM IPAddresses WHERE IPAddress = @Param0", tran?.DBConnection, tran?.DBTransaction, ipBytes);
-                if (reader.Read())
+                using SqliteDataReaderWrapper reader = ExecuteReader("SELECT BanDate, BanEndDate FROM IPAddresses WHERE IPAddress = @Param0", tran?.DBConnection, tran?.DBTransaction, false, ipBytes);
+                if (reader.Reader.Read())
                 {
                     DateTime? banDate = null;
                     DateTime? banEndDate = null;
-                    object val = reader.GetValue(0);
-                    object val2 = reader.GetValue(1);
+                    object val = reader.Reader.GetValue(0);
+                    object val2 = reader.Reader.GetValue(1);
                     if (val != null && val != DBNull.Value)
                     {
                         banDate = ((long)val).ToDateTimeUnixMilliseconds();
@@ -497,30 +497,17 @@ namespace DigitalRuby.IPBanCore
             SqliteDBTransaction tran = transaction as SqliteDBTransaction;
             bool dispose = (tran is null);
             tran ??= BeginTransaction() as SqliteDBTransaction;
-            SqliteDataReader reader;
-
-            // C# can't yield inside a try/catch, so we have to split it up
-            try
-            {
-                // select ip in add pending, remove pending, or remove pending become failed login state
-                reader = ExecuteReader("SELECT IPAddressText, State FROM IPAddresses WHERE State IN (1, 2, 4) ORDER BY IPAddressText", tran.DBConnection, tran.DBTransaction);
-            }
-            catch
-            {
-                RollbackTransaction(tran);
-                throw;
-            }
-
+            using SqliteDataReaderWrapper reader = ExecuteReader("SELECT IPAddressText, State FROM IPAddresses WHERE State IN (1, 2, 4) ORDER BY IPAddressText", tran.DBConnection, tran.DBTransaction, dispose);
             while (true)
             {
                 try
                 {
-                    if (!reader.Read())
+                    if (!reader.Reader.Read())
                     {
                         break;
                     }
-                    ipAddress = reader.GetString(0);
-                    added = (reader.GetInt32(1) == (int)IPAddressState.AddPending);
+                    ipAddress = reader.Reader.GetString(0);
+                    added = (reader.Reader.GetInt32(1) == (int)IPAddressState.AddPending);
                 }
                 catch
                 {
@@ -607,14 +594,14 @@ namespace DigitalRuby.IPBanCore
                 banCutOffUnix = banCutOff.Value.ToUnixMillisecondsLong();
             }
             SqliteDBTransaction tran = transaction as SqliteDBTransaction;
-            using SqliteDataReader reader = ExecuteReader(@"SELECT IPAddressText, LastFailedLogin, FailedLoginCount, BanDate, State, BanEndDate, UserName, Source
+            using SqliteDataReaderWrapper reader = ExecuteReader(@"SELECT IPAddressText, LastFailedLogin, FailedLoginCount, BanDate, State, BanEndDate, UserName, Source
                 FROM IPAddresses
                 WHERE (@Param0 IS NULL AND @Param1 IS NULL) OR (@Param0 IS NOT NULL AND State = 3 AND LastFailedLogin <= @Param0) OR (@Param1 IS NOT NULL AND State IN (0, 1) AND BanEndDate <= @Param1)
                 ORDER BY IPAddress",
-                tran?.DBConnection, tran?.DBTransaction, failLoginCutOffUnix, banCutOffUnix);
-            while (reader.Read())
+                tran?.DBConnection, tran?.DBTransaction, false, failLoginCutOffUnix, banCutOffUnix);
+            while (reader.Reader.Read())
             {
-                yield return ParseIPAddressEntry(reader);
+                yield return ParseIPAddressEntry(reader.Reader);
             }
         }
 
@@ -624,10 +611,10 @@ namespace DigitalRuby.IPBanCore
         /// <returns>IP addresses with non-null ban dates</returns>
         public IEnumerable<string> EnumerateBannedIPAddresses()
         {
-            using SqliteDataReader reader = ExecuteReader("SELECT IPAddressText FROM IPAddresses WHERE BanDate IS NOT NULL AND State = 0 ORDER BY IPAddress", null, null);
-            while (reader.Read())
+            using SqliteDataReaderWrapper reader = ExecuteReader("SELECT IPAddressText FROM IPAddresses WHERE BanDate IS NOT NULL AND State = 0 ORDER BY IPAddress", null, null, false);
+            while (reader.Reader.Read())
             {
-                yield return reader.GetString(0);// ParseIPAddressEntry(reader);
+                yield return reader.Reader.GetString(0);// ParseIPAddressEntry(reader);
             }
         }
 
@@ -671,11 +658,12 @@ namespace DigitalRuby.IPBanCore
         {
             byte[] start = range.Begin.GetAddressBytes();
             byte[] end = range.End.GetAddressBytes();
-            using SqliteDataReader reader = ExecuteReader("SELECT IPAddressText FROM IPAddresses WHERE IPAddress BETWEEN @Param0 AND @Param1 AND length(IPAddress) = length(@Param0) AND length(IPAddress) = length(@Param1); " +
-                "DELETE FROM IPAddresses WHERE IPAddress BETWEEN @Param0 AND @Param1 AND length(IPAddress) = length(@Param0) AND length(IPAddress) = length(@Param1);", null, null, start, end);
-            while (reader.Read())
+            using SqliteDataReaderWrapper reader = ExecuteReader("SELECT IPAddressText FROM IPAddresses WHERE IPAddress BETWEEN @Param0 AND @Param1 AND length(IPAddress) = length(@Param0) AND length(IPAddress) = length(@Param1); " +
+                "DELETE FROM IPAddresses WHERE IPAddress BETWEEN @Param0 AND @Param1 AND length(IPAddress) = length(@Param0) AND length(IPAddress) = length(@Param1);", null, null, false,
+                start, end);
+            while (reader.Reader.Read())
             {
-                yield return reader.GetString(0);
+                yield return reader.Reader.GetString(0);
             }
         }
 
