@@ -40,6 +40,7 @@ namespace DigitalRuby.IPBanCore
     /// </summary>
     [RequiredOperatingSystemAttribute(null, Priority = -99)] // low priority, basically any other firewall is preferred unless this one is explicitly specified in the config
     [CustomName("Memory")]
+    [System.Diagnostics.CodeAnalysis.DynamicallyAccessedMembers(System.Diagnostics.CodeAnalysis.DynamicallyAccessedMemberTypes.All)]
     public class IPBanMemoryFirewall : IPBanBaseFirewall
     {
         public interface IMemoryFirewallRuleRanges
@@ -89,7 +90,7 @@ namespace DigitalRuby.IPBanCore
 
             private readonly List<IPV4Range> ipv4 = new();
             private readonly List<IPV6Range> ipv6 = new();
-            private readonly List<PortRange> portRanges;
+            private readonly PortRange[] portRanges;
 
             public IEnumerable<string> IPV4 => ipv4.Select(r => r.ToIPAddressRange().ToString());
             public IEnumerable<string> IPV6 => ipv6.Select(r => r.ToIPAddressRange().ToString());
@@ -111,17 +112,11 @@ namespace DigitalRuby.IPBanCore
                     // optimized storage, no pointers or other overhead
                     if (range.Begin.AddressFamily == System.Net.Sockets.AddressFamily.InterNetwork)
                     {
-                        uint begin = range.Begin.ToUInt32();
-                        uint end = range.End.ToUInt32();
-                        Debug.Assert(end >= begin);
-                        ipv4.Add(new IPV4Range { Begin = begin, End = end });
+                        ipv4.Add(new IPV4Range(range));
                     }
                     else
                     {
-                        UInt128 begin = range.Begin.ToUInt128();
-                        UInt128 end = range.End.ToUInt128();
-                        Debug.Assert(end.CompareTo(begin) >= 0);
-                        ipv6.Add(new IPV6Range { Begin = begin, End = end });
+                        ipv6.Add(new IPV6Range(range));
                     }
                 }
                 ipv4.TrimExcess();
@@ -129,11 +124,11 @@ namespace DigitalRuby.IPBanCore
                 if (block)
                 {
                     string portString = IPBanFirewallUtility.GetBlockPortRangeString(allowedPorts);
-                    this.portRanges = (string.IsNullOrWhiteSpace(portString) ? new List<PortRange>(0) : portString.Split(',').Select(s => PortRange.Parse(s)).ToList());
+                    this.portRanges = (string.IsNullOrWhiteSpace(portString) ? Array.Empty<PortRange>() : portString.Split(',').Select(s => PortRange.Parse(s)).ToArray());
                 }
                 else
                 {
-                    this.portRanges = allowedPorts;
+                    this.portRanges = allowedPorts.ToArray();
                 }
             }
 
@@ -148,32 +143,37 @@ namespace DigitalRuby.IPBanCore
 
             public bool Contains(uint ipAddress, int port)
             {
-                if (port > -1)
+                bool foundPort = port < 0 || portRanges.Length == 0;
+                if (!foundPort)
                 {
                     foreach (PortRange range in portRanges)
                     {
-                        if (port >= range.MinPort && port <= range.MaxPort)
+                        if (foundPort = (port >= range.MinPort && port <= range.MaxPort))
                         {
-                            return false;
+                            break;
                         }
                     }
                 }
-                return (ipv4.BinarySearch(new IPV4Range { Begin = ipAddress, End = ipAddress }, this) >= 0);
+                return (foundPort && ipv4.BinarySearch(new IPV4Range(ipAddress, ipAddress), this) >= 0);
             }
 
             public bool Contains(UInt128 ipAddress, int port)
             {
-                if (port > -1)
+                bool foundPort = port < 0 || portRanges.Length == 0;
+                if (!foundPort)
                 {
                     foreach (PortRange range in portRanges)
                     {
                         if (port >= range.MinPort && port <= range.MaxPort)
                         {
-                            return false;
+                            if (foundPort = (port >= range.MinPort && port <= range.MaxPort))
+                            {
+                                break;
+                            }
                         }
                     }
                 }
-                return (ipv6.BinarySearch(new IPV6Range { Begin = ipAddress, End = ipAddress }, this) >= 0);
+                return (foundPort && ipv6.BinarySearch(new IPV6Range(ipAddress, ipAddress), this) >= 0);
             }
 
             public IEnumerable<IPAddressRange> EnumerateIPAddressesRanges()
@@ -635,12 +635,14 @@ namespace DigitalRuby.IPBanCore
             return IsIPAddressBlocked(ipAddress, out _, port);
         }
 
-        public bool IsIPAddressBlocked(System.Net.IPAddress ipAddressObj, out string ruleName, int port = -1)
+        public bool IsIPAddressBlocked(System.Net.IPAddress ipAddressObj, out string ruleName, out bool allowed, int port = -1)
         {
+            allowed = false;
             lock (this)
             {
                 if (IsIPAddressAllowed(ipAddressObj, out ruleName, port))
                 {
+                    allowed = true;
                     return false;
                 }
                 else if (ipAddressObj.AddressFamily == System.Net.Sockets.AddressFamily.InterNetwork)
@@ -695,7 +697,7 @@ namespace DigitalRuby.IPBanCore
                 ruleName = null;
                 return false;
             }
-            return IsIPAddressBlocked(ipAddressObj, out ruleName, port);
+            return IsIPAddressBlocked(ipAddressObj, out ruleName, out _, port);
         }
 
         public override void Truncate()
