@@ -238,7 +238,7 @@ namespace DigitalRuby.IPBanCore
                 {
                     try
                     {
-                        domainName = System.DirectoryServices.ActiveDirectory.Domain.GetComputerDomain().Name;
+                        domainName = System.Net.NetworkInformation.IPGlobalProperties.GetIPGlobalProperties().DomainName;
                     }
                     catch
                     {
@@ -257,6 +257,8 @@ namespace DigitalRuby.IPBanCore
                 {
                     FQDN = serverName;
                 }
+
+                Logger.Info("FQDN: {0}", FQDN);
             }
 
             if (string.IsNullOrWhiteSpace(LocalIPAddressString))
@@ -383,7 +385,7 @@ namespace DigitalRuby.IPBanCore
                                         await IPBanDelegate.LoginAttemptFailed(ipAddress, source, userName, MachineGuid, OSName, OSVersion, incrementCount, UtcNow);
                                     }
                                     AddBannedIPAddress(ipAddress, source, userName, bannedIpAddresses, now,
-                                        configBlacklisted, newCount, string.Empty, transaction, failedLogin.External);
+                                        configBlacklisted, newCount, failedLogin.ExtraInfo, transaction, failedLogin.External);
                                 }
                             }
                             else
@@ -519,7 +521,7 @@ namespace DigitalRuby.IPBanCore
                 firewallNeedsBlockedIPAddressesUpdate = true;
             }
 
-            Logger.Warn(startBanDate, "Banning ip address: {0}, user name: {1}, config black listed: {2}, count: {3}, extra info: {4}, duration: {5}",
+            Logger.Warn(startBanDate, "Banning ip address: {0}, user name: {1}, config blacklisted: {2}, count: {3}, extra info: {4}, duration: {5}",
                 ipAddress, userName, configBlacklisted, counter, extraInfo, banTime);
 
             // if this is a delegate callback (counter of 0), exit out - we don't want to run handlers or processes for shared banned ip addresses
@@ -556,52 +558,55 @@ namespace DigitalRuby.IPBanCore
         /// Execute unban process in background
         /// </summary>
         /// <param name="programToRun">Program to run</param>
-        /// <param name="bannedIPAddresses">IP addresses, should be a non-shared collection</param>
-        private void ExecuteExternalProcessForIPAddresses(string programToRun, IReadOnlyCollection<IPAddressLogEvent> bannedIPAddresses)
+        /// <param name="ipAddresses">IP addresses, should be a non-shared collection</param>
+        private void ExecuteExternalProcessForIPAddresses(string programToRun, IReadOnlyCollection<IPAddressLogEvent> ipAddresses)
         {
-            if (bannedIPAddresses is null || bannedIPAddresses.Count == 0 || string.IsNullOrWhiteSpace(programToRun))
+            if (ipAddresses is null || ipAddresses.Count == 0 || string.IsNullOrWhiteSpace(programToRun))
             {
                 return;
             }
-            string[] pieces = programToRun.Split('|');
-            if (pieces.Length != 2)
+            foreach (string process in programToRun.Split('\n'))
             {
-                throw new ArgumentException("Invalid config option for process to run: " + programToRun +
-                    " -- should be two strings, | delimited with program and arguments.");
-            }
-
-            RunTask(() =>
-            {
-                string programFullPath = Path.GetFullPath(pieces[0]);
-                string programArgs = pieces[1];
-
-                foreach (var bannedIp in bannedIPAddresses)
+                string[] pieces = process.Trim().Split('|');
+                if (pieces.Length != 2)
                 {
-                    if (bannedIp is null || string.IsNullOrWhiteSpace(bannedIp.IPAddress))
-                    {
-                        continue;
-                    }
-
-                    string replacedArgs = programArgs.Replace("###IPADDRESS###", bannedIp.IPAddress)
-                        .Replace("###SOURCE###", bannedIp.Source ?? string.Empty)
-                        .Replace("###USERNAME###", bannedIp.UserName ?? string.Empty);
-
-                    try
-                    {
-                        ProcessStartInfo psi = new()
-                        {
-                            FileName = programFullPath,
-                            WorkingDirectory = Path.GetDirectoryName(programFullPath),
-                            Arguments = replacedArgs
-                        };
-                        using Process p = Process.Start(psi);
-                    }
-                    catch (Exception ex)
-                    {
-                        Logger.Error(ex, "Failed to execute process {0} {1}", programFullPath, replacedArgs);
-                    }
+                    throw new ArgumentException("Invalid config option for process to run: " + programToRun +
+                        " -- should be two strings, | delimited with program and arguments.");
                 }
-            });
+
+                RunTask(() =>
+                {
+                    string programFullPath = Path.GetFullPath(pieces[0]);
+                    string programArgs = pieces[1];
+
+                    foreach (var bannedIp in ipAddresses)
+                    {
+                        if (string.IsNullOrWhiteSpace(bannedIp?.IPAddress))
+                        {
+                            continue;
+                        }
+
+                        string replacedArgs = programArgs.Replace("###IPADDRESS###", bannedIp.IPAddress)
+                            .Replace("###SOURCE###", bannedIp.Source ?? string.Empty)
+                            .Replace("###USERNAME###", bannedIp.UserName ?? string.Empty);
+
+                        try
+                        {
+                            ProcessStartInfo psi = new()
+                            {
+                                FileName = programFullPath,
+                                WorkingDirectory = Path.GetDirectoryName(programFullPath),
+                                Arguments = replacedArgs
+                            };
+                            using Process p = Process.Start(psi);
+                        }
+                        catch (Exception ex)
+                        {
+                            Logger.Error(ex, "Failed to execute process {0} {1}", programFullPath, replacedArgs);
+                        }
+                    }
+                });
+            }
         }
 
         private void UpdateBannedIPAddressesOnStart()
@@ -652,7 +657,7 @@ namespace DigitalRuby.IPBanCore
         private void LoadFirewall(IPBanConfig oldConfig)
         {
             IIPBanFirewall existing = Firewall;
-            Firewall = FirewallCreator.CreateFirewall(Config, Firewall);
+            Firewall = FirewallCreator.CreateFirewall(FirewallTypes, Config, Firewall);
             if (existing != Firewall)
             {
                 AddUpdater(Firewall);
@@ -1174,7 +1179,7 @@ namespace DigitalRuby.IPBanCore
                             {
                                 // make sure the ip address is ban pending
                                 AddBannedIPAddress(evt.IPAddress, evt.Source, evt.UserName, bannedIPs,
-                                evt.Timestamp, false, evt.Count, string.Empty, transaction, evt.External);
+                                evt.Timestamp, false, evt.Count, evt.ExtraInfo, transaction, evt.External);
                             }
                             break;
 
