@@ -29,7 +29,6 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
-using System.Management;
 using System.Reflection;
 using System.Runtime.InteropServices;
 using System.Text;
@@ -218,7 +217,7 @@ namespace DigitalRuby.IPBanCore
                 int pos = friendlyName.IndexOf(' ');
                 if (pos > 0)
                 {
-                    string firstWord = friendlyName.Substring(0, pos);
+                    string firstWord = friendlyName[..pos];
 
                     // as long as there are no extended chars, prepend Microsoft prefix
                     // some os will prepend Microsoft in another language
@@ -248,19 +247,66 @@ namespace DigitalRuby.IPBanCore
             catch { return ""; }
         }
 
-        /* WMI can hang/crash the process, especially after Windows updates, don't use for now
-        private static string GetFriendlyNameFromWmi()
+        private static void PopulateUsersWindows(Dictionary<string, bool> newUsers)
         {
-            string result = string.Empty;
-            ManagementObjectSearcher searcher = new ManagementObjectSearcher("SELECT Caption FROM Win32_OperatingSystem");
-            foreach (ManagementObject os in searcher.Get())
+            // Windows: WMIC
+            // wmic useraccount get disabled,name
+            // FALSE username
+            // TRUE  disabledusername
+            string output = StartProcessAndWait("wmic", "useraccount get disabled,name");
+            string[] lines = output.Split('\n').Skip(1).ToArray();
+            foreach (string line in lines)
             {
-                result = os["Caption"].ToString();
-                break;
+                string trimmedLine = line.Trim();
+                int pos = trimmedLine.IndexOf(' ');
+                if (pos >= 0)
+                {
+                    string disabled = trimmedLine[..pos].Trim();
+                    string foundUserName = trimmedLine[pos..].Trim();
+                    _ = bool.TryParse(disabled, out bool disabledBool);
+                    newUsers[foundUserName] = !disabledBool;
+                }
             }
-            return result;
         }
-        */
+
+        private static void PopulateUsersLinux(Dictionary<string, bool> newUsers)
+        {
+            // Linux: /etc/passwd
+            if (File.Exists("/etc/passwd") &&
+                File.Exists("/etc/shadow"))
+            {
+                // enabled users must have an entry in password hash file
+                string[] lines = File.ReadAllLines("/etc/shadow");
+
+                // example line:
+                // root:!$1$Fp$SSSuo3L.xA5s/kMEEIloU1:18049:0:99999:7:::
+                foreach (string[] pieces in lines.Select(l => l.Split(':')).Where(p => p.Length == 9))
+                {
+                    string checkUserName = pieces[0].Trim();
+                    string pwdHash = pieces[1].Trim();
+                    bool hasPwdHash = (pwdHash.Length != 0 && pwdHash[0] != '*' && pwdHash[0] != '!');
+                    newUsers[checkUserName] = hasPwdHash;
+                }
+
+                // filter out nologin users
+                lines = File.ReadAllLines("/etc/passwd");
+
+                // example line:
+                // root:x:0:0:root:/root:/bin/bash
+                foreach (string[] pieces in lines.Select(l => l.Split(':')).Where(p => p.Length == 7))
+                {
+                    // x means shadow file is where the password is at
+                    string checkUserName = pieces[0].Trim();
+                    string nologin = pieces[6];
+                    bool cannotLogin = (nologin.Contains("nologin", StringComparison.OrdinalIgnoreCase) ||
+                        nologin.Contains("/bin/false", StringComparison.OrdinalIgnoreCase));
+                    if (cannotLogin)
+                    {
+                        newUsers[checkUserName] = false;
+                    }
+                }
+            }
+        }
 
         /// <summary>
         /// Get a string representing the operating system
@@ -324,65 +370,50 @@ namespace DigitalRuby.IPBanCore
         /// <returns>Output</returns>
         /// <exception cref="ApplicationException">Exit code did not match allowed exit codes</exception>
         public static string StartProcessAndWait(int timeoutMilliseconds, string program, string args,
-        out int exitCode, params int[] allowedExitCodes)
+            out int exitCode, params int[] allowedExitCodes)
         {
             StringBuilder output = new();
             int _exitCode = -1;
             Exception _ex = null;
             Thread thread = new(new ParameterizedThreadStart((_state) =>
             {
-                Logger.Info($"Executing process {program} {args}...");
-
-                using var process = new Process
+                try
                 {
-                    StartInfo = new ProcessStartInfo(program, args)
+                    Logger.Info($"Executing process {program} {args}...");
+
+                    var startInfo = new ProcessStartInfo(program, args)
                     {
                         CreateNoWindow = true,
                         UseShellExecute = false,
                         WindowStyle = ProcessWindowStyle.Hidden,
-                        RedirectStandardOutput = true,
                         RedirectStandardError = true,
+                        RedirectStandardOutput = true,
                         Verb = processVerb
-                    }
-                };
+                    };
+                    using var process = new Process
+                    {
+                        StartInfo = startInfo
+                    };
 
-                process.OutputDataReceived += (object sender, DataReceivedEventArgs e) =>
-                {
-                    if (!string.IsNullOrEmpty(e.Data))
+                    process.Start();
+                    if (!process.WaitForExit(timeoutMilliseconds))
                     {
-                        lock (output)
-                        {
-                            output.Append("[OUT]: ");
-                            output.AppendLine(e.Data);
-                        }
+                        output.Append("Terminating process due to timeout");
+                        process.Kill();
                     }
-                };
-                process.ErrorDataReceived += (object sender, DataReceivedEventArgs e) =>
-                {
-                    if (!string.IsNullOrEmpty(e.Data))
+                    string stdOut = process.StandardOutput.ReadToEnd();
+                    string stdErr = process.StandardError.ReadToEnd();
+                    output.Append(stdOut);
+                    output.Append(stdErr);
+                    _exitCode = process.ExitCode;
+                    if (allowedExitCodes.Length != 0 && Array.IndexOf(allowedExitCodes, process.ExitCode) < 0)
                     {
-                        lock (output)
-                        {
-                            output.Append("[ERR]: ");
-                            output.AppendLine(e.Data);
-                        }
+                        _ex = new ApplicationException($"Program {program} {args}: failed with exit code {process.ExitCode}, output: {output}");
                     }
-                };
-                process.Start();
-                process.BeginOutputReadLine();
-                process.BeginErrorReadLine();
-                if (!process.WaitForExit(timeoutMilliseconds))
-                {
-                    lock (output)
-                    {
-                        output.Append("[ERR]: Terminating process due to timeout");
-                    }
-                    process.Kill();
                 }
-                _exitCode = process.ExitCode;
-                if (allowedExitCodes.Length != 0 && Array.IndexOf(allowedExitCodes, process.ExitCode) < 0)
+                catch (Exception ex)
                 {
-                    _ex = new ApplicationException($"Program {program} {args}: failed with exit code {process.ExitCode}, output: {output}");
+                    _ex = ex;
                 }
             }));
             thread.Start();
@@ -399,7 +430,7 @@ namespace DigitalRuby.IPBanCore
             return output.ToString();
         }
 
-        private static Dictionary<string, bool> users = new(StringComparer.OrdinalIgnoreCase);
+        private static Dictionary<string, bool> users = new(StringComparer.OrdinalIgnoreCase); // user name, enabled
         private static DateTime usersExpire = IPBanService.UtcNow;
 
         /// <summary>
@@ -422,96 +453,39 @@ namespace DigitalRuby.IPBanCore
             userName = userName.Trim();
 
             // check cache first
-            bool enabled;
-            if (usersExpire > IPBanService.UtcNow)
+            bool cacheExpired = (usersExpire <= IPBanService.UtcNow);
+            if (cacheExpired)
             {
-                users.TryGetValue(userName, out enabled);
-                return enabled;
-            }
-
-            try
-            {
-                if (isWindows)
+                try
                 {
-                    // Windows: WMI
-                    if (usersExpire <= IPBanService.UtcNow)
+                    Dictionary<string, bool> newUsers = new(StringComparer.OrdinalIgnoreCase);
+                    lock (users)
                     {
-                        lock (users)
+                        cacheExpired = (usersExpire <= IPBanService.UtcNow);
+                        if (cacheExpired)
                         {
-                            Dictionary<string, bool> newUsers = new(StringComparer.OrdinalIgnoreCase);
-                            SelectQuery query = new("Win32_UserAccount");
-                            ManagementObjectSearcher searcher = new(query);
-                            foreach (ManagementObject user in searcher.Get())
+                            if (isWindows)
                             {
-#pragma warning disable CA1507 // Use nameof to express symbol names
-                                string foundUserName = user["Name"]?.ToString();
-#pragma warning restore CA1507 // Use nameof to express symbol names
-                                if (!string.IsNullOrWhiteSpace(foundUserName))
-                                {
-                                    newUsers[foundUserName] = user["Disabled"] is null || user["Disabled"].Equals(false);
-                                }
+                                PopulateUsersWindows(newUsers);
                             }
+                            else if (isLinux)
+                            {
+                                PopulateUsersLinux(newUsers);
+                            }
+                            // TODO: MAC
+
                             usersExpire = IPBanService.UtcNow + UserIsActiveCacheTime;
                             users = newUsers;
                         }
                     }
                 }
-                else if (isLinux)
+                catch (Exception ex)
                 {
-                    // Linux: /etc/passwd
-                    if (File.Exists("/etc/passwd") && File.Exists("/etc/shadow"))
-                    {
-                        // check for cache expire and refill if needed
-                        if (usersExpire <= IPBanService.UtcNow)
-                        {
-                            lock (users)
-                            {
-                                Dictionary<string, bool> newUsers = new(StringComparer.OrdinalIgnoreCase);
-
-                                // enabled users must have an entry in password hash file
-                                string[] lines = File.ReadAllLines("/etc/shadow");
-
-                                // example line:
-                                // root:!$1$Fp$SSSuo3L.xA5s/kMEEIloU1:18049:0:99999:7:::
-                                foreach (string[] pieces in lines.Select(l => l.Split(':')).Where(p => p.Length == 9))
-                                {
-                                    string checkUserName = pieces[0].Trim();
-                                    string pwdHash = pieces[1].Trim();
-                                    bool hasPwdHash = (pwdHash.Length != 0 && pwdHash[0] != '*' && pwdHash[0] != '!');
-                                    newUsers[checkUserName] = hasPwdHash;
-                                }
-
-                                // filter out nologin users
-                                lines = File.ReadAllLines("/etc/passwd");
-
-                                // example line:
-                                // root:x:0:0:root:/root:/bin/bash
-                                foreach (string[] pieces in lines.Select(l => l.Split(':')).Where(p => p.Length == 7))
-                                {
-                                    // x means shadow file is where the password is at
-                                    string checkUserName = pieces[0].Trim();
-                                    string nologin = pieces[6];
-                                    bool cannotLogin = (nologin.Contains("nologin", StringComparison.OrdinalIgnoreCase) ||
-                                        nologin.Contains("/bin/false", StringComparison.OrdinalIgnoreCase));
-                                    if (cannotLogin)
-                                    {
-                                        newUsers[checkUserName] = false;
-                                    }
-                                }
-                                usersExpire = IPBanService.UtcNow + UserIsActiveCacheTime;
-                                users = newUsers;
-                            }
-                        }
-                    }
+                    Logger.Error("Error determining if user is active", ex);
                 }
-                // TODO: MAC
-            }
-            catch (Exception ex)
-            {
-                Logger.Error("Error determining if user is active", ex);
             }
 
-            users.TryGetValue(userName, out enabled);
+            users.TryGetValue(userName, out bool enabled);
             return enabled;
         }
 
